+++ serve.go
@@ -29,7 +29,7 @@ import (
 
 // Increment and remove "+" in release commits.
 // Add "+" after release commits.
-const version = "v0.2.0"
+const version = "v0.2.1"
 
 // Flags in alphabetical order, just like "-h" prints them
 var (